backend/requirements.txt
passlib~=1.7.4
fastapi~=0.115.6
SQLAlchemy~=2.0.36
python-dotenv~=1.0.1
uvicorn~=0.32.1
sqlalchemy-repr==0.1.0
psycopg2==2.9.10